MS-8 gauss jordan

+++ src/matrix/matrix_unit-test.h
@@ -12,16 +12,17 @@
 
 PROTOTYPE_FOR_UNIT_TEST(void swap_row(matrix_t* A, unsigned i, unsigned j));
 
-PROTOTYPE_FOR_UNIT_TEST(void matrix_set(struct matrix* A, const float* e));
+PROTOTYPE_FOR_UNIT_TEST(void matrix_set(struct matrix* A,
+                                        const long double* e));
 
 PROTOTYPE_FOR_UNIT_TEST(bool matrix_eq(const struct matrix* A,
                                        const struct matrix* B));
 
 PROTOTYPE_FOR_UNIT_TEST(void mult_row(struct matrix* A,
                                       unsigned i,
-                                      float lambda));
+                                      long double lambda));
 
 PROTOTYPE_FOR_UNIT_TEST(
-  void add_row(struct matrix* A, unsigned i, unsigned j, float lambda));
+  void add_row(struct matrix* A, unsigned i, unsigned j, long double lambda));
 
 #endif /* !__MATRIX_UNIT_TEST_H__ */
